# api/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IOrder": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"ready": {
						"type": "boolean"
					},
					"userName": {
						"type": "string"
					},
					"userMob": {
						"type": "string"
					},
					"orderItems": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"total",
					"ready",
					"userName",
					"userMob",
					"orderItems"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICartItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICart": {
				"properties": {
					"mains": {
						"items": {
							"$ref": "#/components/schemas/ICartItem"
						},
						"type": "array"
					},
					"drinks": {
						"items": {
							"$ref": "#/components/schemas/ICartItem"
						},
						"type": "array"
					},
					"desserts": {
						"items": {
							"$ref": "#/components/schemas/ICartItem"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/menu/getItems": {
			"get": {
				"operationId": "GetItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Menu"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/menu/getCategories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Menu"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/menu/generateMenu": {
			"post": {
				"operationId": "GenerateMenu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Menu"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/order/create": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"order": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"order",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrder"
							}
						}
					}
				}
			}
		},
		"/api/order/update": {
			"post": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"properties": {
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"id"
										],
										"type": "object"
									},
									{
										"$ref": "#/components/schemas/IOrder"
									}
								]
							}
						}
					}
				}
			}
		},
		"/api/order/checkOffer": {
			"post": {
				"operationId": "CheckOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"amountAfterDiscout": {
											"type": "string"
										},
										"totalDiscount": {
											"type": "string"
										},
										"totalAmount": {
											"type": "number",
											"format": "double"
										},
										"cart": {
											"$ref": "#/components/schemas/ICart"
										},
										"message": {
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									},
									"required": [
										"amountAfterDiscout",
										"totalDiscount",
										"totalAmount",
										"cart",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICart"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/get-all": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user-permission/create": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"roles": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"roles",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/update/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"roles": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"roles",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/delete/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}